requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
fastapi==0.109.2
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
# psycopg2-binary==2.9.3
psycopg2==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlmodel==0.0.14
starlette==0.36.3
strawberry-graphql @ git+https://github.com/strawberry-graphql/strawberry.git@c926395ff96acd22c2fd0170edc308f8548e288e
typing_extensions==4.9.0
uvicorn==0.27.0.post1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.7